replace prettier-plugin-import-sort with prettier-plugin-organize-imports
The main motivation is to remove transient depedencies against
older version of typescript

prettier-plugin-import-sort depends on import-sort-parser-typescript
which itself still depends on typescript^3.2.4, resulting in both
v3 and v5 of typescript being installed side-by-side and causing
issue in some environment

prettier-plugin-organize-imports does not support any customization
on how the import is organized. Instead it makes use of the built-in
organizeImports feature of the typescript language service API.
The same API that is used by "Organize Imports" in VS Code.

while the ability to customize the way import is organized is "nice",
arguably, it is better to get closer to how the "default" behavior are,
since there are less surprises